split out request building and request auth
